requirements.txt
alabaster==1.0.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
blis==0.7.11
cachelib==0.13.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
cryptography==44.0.2
cymem==2.0.11
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docutils==0.21.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
executing==2.2.0
ExifRead==3.0.0
faiss-cpu==1.10.0
fastjsonschema==2.21.1
filelock==3.18.0
filetype==1.2.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
fqdn==1.5.1
fsspec==2025.3.0
google-api-core==2.24.2
google-api-python-client==2.122.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.69.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.8
ImageHash==4.3.2
imagesize==1.4.1
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.9.0
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
langcodes==3.5.0
language_data==1.3.0
lxml==5.3.1
Mako==1.3.9
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
msgspec==0.19.0
murmurhash==1.0.12
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.8.1
notebook==7.3.3
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
openai==1.12.0
opencv-python==4.11.0.86
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pdf2image==1.16.3
pexpect==4.9.0
pgvector==0.4.0
pillow==10.2.0
platformdirs==4.3.7
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==6.30.2
psutil==7.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
PyMuPDF==1.25.4
pyparsing==3.2.3
PyPDF2==3.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.0
python-json-logger==3.3.0
python-magic==0.4.27
pytz==2025.1
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.4.0
scipy==1.15.2
selenium==4.18.1
Send2Trash==1.8.3
sentence-transformers==2.5.1
setuptools==77.0.3
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==8.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.27
srsly==2.5.1
stack-data==0.6.3
sympy==1.13.1
terminado==0.18.1
thinc==8.2.5
threadpoolctl==3.6.0
tiktoken==0.6.0
tinycss2==1.4.0
tokenizers==0.15.2
torch==2.2.1
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.37.2
trio==0.29.0
trio-websocket==0.12.2
typer==0.9.4
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
Wand==0.6.13
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.3.4
webcolors==24.11.1
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.1
widgetsnbextension==4.0.13
wsproto==1.2.0
WTForms==3.2.1
pytest==8.0.0
pytest-cov==4.1.0
tenacity==8.2.3
redis==5.0.1
